Update minidata_mutation_calling_v3.wdl
This is a working wdl. It scatters across the tumor samples. It also scatters across chromosomes for just the ApplyBaseRecalibrator task. Two additional tasks had to be added in order scatter across chromosomes and then gather. The code possibly needs some input. But it runs and works and calls the mutations